# pkg/url_shotener_pb/api/url_shortener/url_shotener.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/url_shortener/url_shotener.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "UrlShortenerService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/url/create": {
      "get": {
        "summary": "Create short url",
        "operationId": "UrlShortenerService_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/url_shortenerCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "UrlShortenerService"
        ]
      }
    },
    "/url/list": {
      "get": {
        "summary": "Get all urls with short",
        "operationId": "UrlShortenerService_GetList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/url_shortenerGetListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "UrlShortenerService"
        ]
      }
    },
    "/url/remove": {
      "post": {
        "summary": "Remove short url",
        "operationId": "UrlShortenerService_Remove",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/url_shortenerRemoveResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/url_shortenerRemoveRequest"
            }
          }
        ],
        "tags": [
          "UrlShortenerService"
        ]
      }
    },
    "/url/{url}": {
      "get": {
        "summary": "Get full url by short",
        "operationId": "UrlShortenerService_GetFullURL",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/url_shortenerGetFullURLResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "url",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "UrlShortenerService"
        ]
      }
    }
  },
  "definitions": {
    "GetListResponseInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "url": {
          "type": "string"
        },
        "shortUrl": {
          "type": "string"
        },
        "redirected": {
          "type": "integer",
          "format": "int32"
        },
        "ctreatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "url_shortenerCreateResponse": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string"
        }
      }
    },
    "url_shortenerGetFullURLResponse": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string"
        }
      }
    },
    "url_shortenerGetListResponse": {
      "type": "object",
      "properties": {
        "urls": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GetListResponseInfo"
          }
        }
      }
    },
    "url_shortenerRemoveRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "url_shortenerRemoveResponse": {
      "type": "object"
    }
  }
}